test: add assertions for coordinate transform
