--- tests/unit/test_coordinate_fix.js
@@ -1,106 +1,0 @@
-/**
- * Mouse Coordinate Fix Test Script
- * Tests the viewport-to-canvas coordinate transformation fix
- */
-
-// Test the coordinate transformation with mock data
-function testCoordinateTransformation() {
-    console.log('🧪 Testing Mouse Coordinate Transformation Fix\n');
-    
-    // Mock canvas element with bounding rectangle
-    const mockCanvas = {
-        getBoundingClientRect: () => ({
-            left: 50,   // Canvas is 50px from left edge
-            top: 100,   // Canvas is 100px from top edge (after scroll)
-            width: 800,
-            height: 600
-        })
-    };
-    
-    // Mock camera object
-    const mockCamera = {
-        scale: 1,
-        x: 0,
-        y: 0,
-        
-        // Old broken method (viewport coordinates only)
-        screenToWorldOld: function(screenX, screenY) {
-            return {
-                x: (screenX / this.scale) + this.x,
-                y: (screenY / this.scale) + this.y
-            };
-        },
-        
-        // New fixed method (viewport-to-canvas conversion)
-        screenToWorld: function(screenX, screenY, canvasElement = null) {
-            let canvasX = screenX;
-            let canvasY = screenY;
-            
-            // Convert viewport coordinates to canvas-relative coordinates if canvas element provided
-            if (canvasElement) {
-                const rect = canvasElement.getBoundingClientRect();
-                canvasX = screenX - rect.left;
-                canvasY = screenY - rect.top;
-            }
-            
-            return {
-                x: (canvasX / this.scale) + this.x,
-                y: (canvasY / this.scale) + this.y
-            };
-        }
-    };
-    
-    // Test scenarios
-    const testCases = [
-        {
-            name: "Mouse click at viewport position (300, 200)",
-            viewportX: 300,
-            viewportY: 200,
-            description: "User clicks mouse at position 300, 200 on screen"
-        },
-        {
-            name: "Mouse click at viewport position (150, 150) after scroll",
-            viewportX: 150,
-            viewportY: 150,
-            description: "User clicks mouse at position 150, 150 after scrolling page"
-        }
-    ];
-    
-    testCases.forEach((testCase, index) => {
-        console.log(`\n=== Test Case ${index + 1}: ${testCase.name} ===`);
-        console.log(`Description: ${testCase.description}`);
-        console.log(`Canvas position: left=${mockCanvas.getBoundingClientRect().left}px, top=${mockCanvas.getBoundingClientRect().top}px`);
-        
-        // Test old broken method
-        const oldResult = mockCamera.screenToWorldOld(testCase.viewportX, testCase.viewportY);
-        console.log(`❌ OLD (Broken): Viewport (${testCase.viewportX}, ${testCase.viewportY}) → World (${oldResult.x}, ${oldResult.y})`);
-        
-        // Test new fixed method
-        const newResult = mockCamera.screenToWorld(testCase.viewportX, testCase.viewportY, mockCanvas);
-        console.log(`✅ NEW (Fixed):  Viewport (${testCase.viewportX}, ${testCase.viewportY}) → Canvas (${testCase.viewportX - mockCanvas.getBoundingClientRect().left}, ${testCase.viewportY - mockCanvas.getBoundingClientRect().top}) → World (${newResult.x}, ${newResult.y})`);
-        
-        const deltaX = Math.abs(newResult.x - oldResult.x);
-        const deltaY = Math.abs(newResult.y - oldResult.y);
-        console.log(`📏 Coordinate difference: Δx=${deltaX}px, Δy=${deltaY}px`);
-        
-        if (deltaX > 0 || deltaY > 0) {
-            console.log(`🎯 Fix Impact: Selection box will be ${deltaX}px left, ${deltaY}px up from wrong position`);
-        } else {
-            console.log(`⚡ No difference: Canvas is at viewport origin`);
-        }
-    });
-    
-    console.log('\n🔧 Technical Implementation:');
-    console.log('1. Enhanced Camera.screenToWorld(x, y, canvasElement) to accept canvas element');
-    console.log('2. Added viewport-to-canvas coordinate conversion using getBoundingClientRect()');
-    console.log('3. Updated all mouse event handlers in SelectionSystem, InputBatcher, etc.');
-    console.log('4. Selection box now appears exactly where user drags cursor');
-    
-    console.log('\n✅ Mouse Coordinate Desync Issue RESOLVED!');
-    console.log('Selection will work accurately regardless of page scroll position.');
-}
-
-// Run the test
-testCoordinateTransformation();
-
-export { testCoordinateTransformation };
